refactor(sqlite): unify record/token adapter logic behind a SqlExecutor
Follow-up to feedback on the native-adapter PR: record-adapter-sqljs
and record-adapter-sqlite had near-identical method bodies (same SQL
text, same parameter order) for every CRUD/query/version/type/
association/token operation, differing only in each engine's call
convention (sql.js's array-bind-then-step vs. node:sqlite's
spread-args run/get/all). That's now unified.

sqlite-shared adds:
- SqlExecutor: a small interface (exec/run/get/all) normalizing the
  two conventions, plus isForeignKeyViolation (verified identical
  error message on both engines already, just written twice before).
- FtsStrategy: the one place FTS4 and FTS5 genuinely differ (FTS5's
  external-content table needs its special 'delete' command with the
  *old* content to unindex a row; FTS4's plain DELETE is fine either
  way). fts4Strategy/fts5Strategy are the concrete strategies.
- SharedSqlRecordLogic: every StackRecordAdapter method (createRecord,
  getRecord, patchContent, deleteRecord, associate/dissociate,
  versions, types, deleteUnreferencedAttachmentRecords, ...), written
  once against a SqlExecutor + FtsStrategy instead of twice against
  two different database APIs.
- SharedTokenLogic: same treatment for StackTokenStore
  (createToken/lookupToken/listTokens/revokeToken).
- insertConfigRecord/readStackConfig: the _config@1 record read/write,
  also identical across engines modulo the call convention.

Both adapters now implement a thin per-engine SqlExecutor
(SqlJsExecutor, NativeSqliteExecutor) and construct
SharedSqlRecordLogic/SharedTokenLogic with it, an FTS strategy, and an
onWrite hook (sql.js's export+rename persist(); omitted for native,
whose writes are already durable via WAL). What's left in each
package's index.ts is genuinely engine-specific: WASM init vs.
DatabaseSync construction, pragma setup, the FTS4->FTS5 migration, and
lifecycle (flush/close).

Net effect: record-adapter-sqljs's index.ts drops from 676 to 373
lines; record-adapter-sqlite's from 614 to 310 (+ a 23-line executor +
its token-store.ts shrinking similarly). No behavior change — every
existing test in both packages passes completely unmodified (95 +
71 = 166 tests), which is the actual proof this refactor is
behavior-preserving rather than just believed to be.
