+++ packages/sqlite-shared/src/index.ts
@@ -22,6 +22,11 @@ export {
   type DecodedCursor,
 } from './cursor.js';
 export { rowToRecord, rowToAssociation, rowToType, rowToVersion, toMs, fromMs } from './mappers.js';
-export { sanitizeFts4Query } from './fts.js';
-export { sanitizeFts5Query } from './fts5.js';
+export { sanitizeFts4Query, fts4Strategy } from './fts.js';
+export { sanitizeFts5Query, fts5Strategy } from './fts5.js';
+export { type FtsStrategy } from './fts-strategy.js';
 export { acquireLock, releaseLock } from './lock.js';
+export { type SqlExecutor, isForeignKeyViolation } from './executor.js';
+export { insertConfigRecord, readStackConfig, type StackConfig } from './config.js';
+export { SharedSqlRecordLogic, type SharedSqlRecordLogicDeps } from './record-logic.js';
+export { SharedTokenLogic, type SharedTokenLogicDeps } from './token-logic.js';
